Set RUSTDOCFLAGS to avoid issues with linkage for doctests. (#143)
Per issue #142, `cargo afl test` and other commands that would run
doctests would encounter issues with linkage. This commit changes `cargo
afl` to set RUSTDOCFLAGS so rustdoc will be able to find the required
libs. Further options are passed to ensure doctests are built and run in
as similar a fashion to other code. This avoids other gotchas like cfg
fuzzing being set for normal builds, but not for doctests.