Use `OUT_DIR` as a data directory in build scripts
This change addresses the same problem as #185. It turns out that build
scripts are only meant to write to OUT_DIR, and in general it is untidy
to write things all over the place. docs.rs is specially sensitive to
this.

Also, docs.rs does not set the feature flag "docsrs" by itself, but a
Cargo package metadata section can be used to instruct docs.rs to pass
that flag to Cargo.