Remove images from published crates.io source code
While reviewing dependencies I noticed that the published source code of
the `afl` crate includes binary data (images). Removing them reduces the
size of the published tar-ball from 1.3MB to 6KB, which is a size
reduction of several orders of magnitude. This would save crates.io more
than 125GB of traffik per 90 days according to their download numbers.

There are two main benifits of that change:

* Drastical reduction of download size
* Removes binary data from a crates that contains a build script, which
  hopefully makes the crate a bit more robust against potential supply
chain attacks similar to the xz incident.

The images seem to be used in the Readme. I've replaced them with links
to the images hosted at github there so that the readme is still
rendered correctly at crates.io.